+++ web/src/features/score-analytics/components/cards/DistributionCategoricalCard.tsx
@@ -10,8 +10,8 @@ import { Tabs, TabsList, TabsTrigger } from "@/src/components/ui/tabs";
 import { Loader2 } from "lucide-react";
 import { useScoreAnalytics } from "../ScoreAnalyticsProvider";
 import { ScoreDistributionCategoricalChart } from "../charts/ScoreDistributionCategoricalChart";
-import { getScoreCategoryColors } from "../../libs/color-scales";
-import { SamplingDetailsHoverCard } from "../ScoreAnalyticsNoticeBanner";
+import { getScoreCategoryColors } from "../../lib/color-scales";
+import { SamplingDetailsHoverCard } from "../SamplingDetailsHoverCard";
 
 type DistributionTab = "score1" | "score2" | "all" | "matched";
