+++ web/src/features/score-analytics/components/charts/HeatmapCell.tsx
@@ -1,5 +1,5 @@
-import { type HeatmapCell } from "@/src/features/scores/components/score-analytics/libs/heatmap-utils";
-import { getContrastColor } from "@/src/features/scores/components/score-analytics/libs/color-scales";
+import { type HeatmapCell } from "@/src/features/score-analytics/lib/heatmap-utils";
+import { getContrastColor } from "@/src/features/score-analytics/lib/color-scales";
 import {
   Tooltip,
   TooltipContent,
