+++ web/src/features/score-analytics/components/charts/HeatmapPlaceholder.tsx
