+++ web/src/features/score-analytics/components/charts/HeatmapSkeleton.tsx
