+++ web/src/features/score-analytics/components/charts/ObjectTypeFilter.tsx
@@ -5,7 +5,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/src/components/ui/select";
-import { type ObjectType } from "@/src/features/scores/components/score-analytics/libs/analytics-url-state";
+import { type ObjectType } from "@/src/features/score-analytics/lib/analytics-url-state";
 
 const OBJECT_TYPE_OPTIONS: Array<{ value: ObjectType; label: string }> = [
   { value: "all", label: "All Objects" },
