+++ web/src/features/score-analytics/components/charts/ScoreChartLegendContent.tsx
