+++ web/src/features/score-analytics/components/charts/ScoreTimeSeriesChart.tsx
