+++ web/src/features/score-analytics/lib/ScoreChartTooltip.tsx
