+++ web/src/features/score-analytics/lib/chart-formatters.ts
