+++ web/src/features/score-analytics/lib/color-scales.ts
