+++ web/src/features/score-analytics/lib/heatmap-utils.ts
