+++ web/src/features/score-analytics/lib/statistics-utils.ts
